requirements.txt
faster-whisper>=1.1.0,<2.0.0
python-docx>=1.1.2,<2.0.0
numpy>=1.26.0,<2.0.0
soundcard>=0.4.3,<0.5.0
resemblyzer>=0.1.4,<0.2.0
simpleaudio>=1.0.4,<2.0.0
pyinstaller>=6.16.0,<7.0.0
